Deploy rewards to Athens2
